feat(native): when translating OutputNode, properly set rootFragment
Once Velox PR #15404 is merged in, it will be possible to signal Velox
that the particular PartitionedOutputNode will be used only to hold
data for the coordinator as the final output of a plan. This is set to
false by default, however, to properly translate OutputNode, we need to
set this to true.

(cherry picked from commit a18e551573e05d94898bb355c2f20b99a24ed2cc)